refactor(shell-ui): extract the MCP task aggregator into its own module
Review feedback on the background-task plumbing:

- Move the host-side task aggregator (task model, shared registry and the
  single `getTaskStatus` host tool) out of MCPRegistrar.tsx into mcp/tasks.ts.
  MCPRegistrar is back to just registering a federated app's tools.
- Replace the inline `ret instanceof Object && 'taskId' in ret` check plus
  `as Task` cast with a named `TaskHandle` type and an `isTaskHandle()` guard.
  Kept structural rather than a class + `instanceof`: tool modules come from
  other apps' bundles over module federation and shell-ui exposes no runtime
  module to them, so each side would bundle its own class and `instanceof`
  would always be false. The guard is also stricter than before — it requires
  taskId to actually be a string.
- Drop the `as TaskStatus` cast in the unknown-task reply by annotating the
  return type instead, which is what checks the literal against the union.

Also covers the guard with a table test for the results it must not track.
